Added a new file for the updated logic of Schmidt to be done in just O(v+e) logic. Code has a small issue in how to use visiting time in order to prevent the redundant and incorrect inclusion of the extra opposite of back-edges in the back-edges graph. Maybe that resolves the seg fault. Lets see that tomorrow please!
